+++ .dir-locals.el
@@ -7,6 +7,7 @@
                    (elisp-lint-ignored-validators . ("fill-column"))
                    (elisp-lint-indent-specs . ((cider-propertize-region . 1)
                                                (cl-flet . 1)
+                                               (emidje-outline-section . 1)
                                                (if-let . 2)
                                                (magit-define-popup . 1)
                                                (nrepl-dbind-response . 2)